dataingest/requirements.txt
bcrypt==4.1.3
psycopg2-binary==2.9.9
pymysql==1.1.1
pymongo==4.7.2
faker==25.2.0
python-dotenv==1.0.1
pytest==8.2.0
